requirements.txt
asyncio==3.4.3
elevenlabs==1.1.2
openai==1.12.0
pyaudio==0.2.14
pyserial==3.5
python-dotenv==1.0.1
SpeechRecognition==3.10.3